Make DIDJWKSet a private class in SdJwtVcJwtProcessor and expose it via a method of its companion object.
